Fixed issue where media was trying to be deleted that didn't exist.

+++ project2-back/src/main/java/net/revature/project1/service/PostService.java
@@ -220,7 +220,10 @@ public class PostService {
             return PostEnum.UNAUTHORIZED;
         }
 
-        fileService.deleteFile(postToDelete.getMedia());
+        if(postToDelete.getMedia() != null && !postToDelete.getMedia().isEmpty()){
+            fileService.deleteFile(postToDelete.getMedia());
+        }
+
         postRepo.delete(postToDelete);
         return PostEnum.SUCCESS;
     }
